Transfer main payment logic from service to api package
